+++ packages/schema-tree/src/nodes/SchemaTreeNode.tsx
@@ -19,7 +19,7 @@ export const SchemaTreeNode: FC<{
     <BaseTreeNode asChild className={className} nodeObject={nodeObject}>
       <div className="flex w-full items-center space-x-2">
         <FolderIcon size="16px" className="shrink-0 text-yellow-500" />
-        <span className="text-sm">{nodeObject.name}</span>
+        <span>{nodeObject.name}</span>
       </div>
       <TreeNodeActionsMenu>
         <TreeNodeActionsMenuItem