+++ packages/sql-editor/src/SqlEditorSlice.tsx
@@ -17,6 +17,7 @@ import * as arrow from 'apache-arrow';
 import {csvFormat} from 'd3-dsv';
 import {saveAs} from 'file-saver';
 import {produce} from 'immer';
+import {createId} from '@paralleldrive/cuid2';
 
 export type QueryResult =
   | {status: 'loading'; isBeingAborted?: boolean; controller: AbortController}
@@ -156,13 +157,13 @@ export function createSqlEditorSlice<
           const blob = new Blob([csvFormat(results.toArray())], {
             type: 'text/plain;charset=utf-8',
           });
-          saveAs(blob, filename || `export-${genRandomStr(5)}.csv`);
+          saveAs(blob, filename || `export-${createId().substring(0, 5)}.csv`);
         },
 
         createQueryTab: (initialQuery = '') => {
           const sqlEditorConfig = get().config.sqlEditor;
           const newQuery = {
-            id: genRandomStr(8),
+            id: createId(),
             name: generateUniqueName(
               'Untitled',
               sqlEditorConfig.queries.map((q) => q.name),